+++ src/__tests__/exports.ts
@@ -13,6 +13,10 @@ describe("Package Exports", () => {
       "Callout",
       "Camera",
       "UserLocation",
+      "NativeUserLocation",
+
+      // Hooks
+      "useCurrentPosition",
 
       // modules
       "OfflineManager",
@@ -39,7 +43,6 @@ describe("Package Exports", () => {
       "Images",
 
       // Constants
-      "UserLocationRenderMode",
       "StyleURL",
       "StyleSource",
       "OfflinePackDownloadState",
@@ -50,7 +53,6 @@ describe("Package Exports", () => {
       "removeCustomHeader",
 
       "setConnected",
-      "requestAndroidLocationPermissions",
 
       // animated
       "Animated",