+++ src/components/user-location/UserLocationPuckHeading.tsx
@@ -22,11 +22,12 @@ export const UserLocationPuckHeading = memo(
   ({ sourceID, belowLayerID, heading }: UserLocationPuckHeadingProps) => (
     <SymbolLayer
       id="mlrn-user-location-puck-heading"
+      testID="mlrn-user-location-puck-heading"
       sourceID={sourceID}
       belowLayerID={belowLayerID}
       style={{
-        iconRotate: heading,
         ...layerStyle,
+        iconRotate: heading,
       }}
     />
   ),